pyproject.toml
[project]
name = "data-centered-ingestion"
version = "0.1.0"
description = "Automated ingestion pipeline for data-centered.com knowledge base"
readme = "docs/ingestion-pipeline-design.md"
requires-python = ">=3.11"
dependencies = [
    "unstructured[html]>=0.15.0",
    "dspy-ai>=2.5.0",
    "pyyaml>=6.0",
    "python-slugify>=8.0",
    "httpx>=0.27.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5.0",
]
# Full document support (PDF, DOCX, etc.)
full = [
    "unstructured[all-docs]>=0.15.0",
    "feedparser>=6.0",
]

[project.scripts]
ingest = "ingest:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]  # Line length handled by formatter
